refactor: remove beneficiaryVASP and consolidate travel rule tests
